requirements.txt
numpy == 2.0.1
pillow == 10.4.0
svg.py == 1.4.3
